Reorganize schemas and types declarations
